iface: output wifi name if possible
If iw is present on the system, it will try to find the wifi name and display that as short text.
As long text the IP and the wifi name will be shown.
I removed the spaces after the label. This aligns with all other blocks.
I.e. that you don't have to set LABEL=mywifi